updates based on PR feedback

+++ lib/ext/model.rb
@@ -1,6 +1,12 @@
 module FHIR
   class Model
-    attr_reader :client
+    class << self
+      cattr_accessor :client
+    end
+
+    def client
+      @client || self.class.client
+    end
 
     def client=(client)
       @client = client
@@ -15,16 +21,16 @@ module FHIR
       end
     end
 
-    def self.read(client, id)
+    def self.read(id, client = self.client)
       client.read(self, id).resource
     end
 
-    def self.search(client, params = {})
+    def self.search(params = {}, client = self.client)
       client.search(self, search: { parameters: params }).resource
     end
 
-    def self.create(client, model)
-      model = self.new(model) unless model.is_a?(self)
+    def self.create(model, client = self.client)
+      model = new(model) unless model.is_a?(self)
       client.create(model).resource
     end
 
@@ -33,7 +39,16 @@ module FHIR
     end
 
     def destroy
-      client.destroy(self, id)
+      client.destroy(self, id) unless id.nil?
+      nil
+    end
+
+    def save
+      if id.nil?
+        self.class.create(self, client)
+      else
+        update
+      end
     end
   end
 end
